集成旧工程的应用代码，修改cmsis接口应用版本，使用V2接口实现涉及到的功能

+++ BAU/bau_stm32f407vgt6/Core/Inc/mpu6050.h
@@ -12,9 +12,9 @@ extern "C" {
  * 加速度计采样值
 */
 struct MPU6050_ACCEL {
-	int32_t x;
-	int32_t y;
-	int32_t z;
+	float x;
+	float y;
+	float z;
 };
 
 
@@ -22,9 +22,9 @@ struct MPU6050_ACCEL {
  * 陀螺仪采样值
 */
 struct MPU6050_GYRO {
-	int32_t x;
-	int32_t y;
-	int32_t z;
+	float x;
+	float y;
+	float z;
 };
